% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_params.R
\name{findParams}
\alias{findParams}
\title{findParams}
\usage{
findParams(param)
}
\arguments{
\item{param}{Parameter to retrieve}
}
\description{
helper function to retrieve valid values for parameters 
(mostly used internally)
}
\details{
Given the pletoric number of parameters a this helper function is provided. 
Please see the 
\href{https://developers.google.com/youtube/v3/docs/search/list}{official documentation} 
for more information. Note that not all parameters (i.e.: \code{channel.id}) 
cannot, by nature, be tested.

Valid parameters: 
\itemize{
\item \code{order}
\item \code{video.dimension}
\item \code{video.caption}
\item \code{video.duration}
\item \code{video.definition}
\item \code{video.embeddable}
\item \code{video.syndicated}
\item \code{video.license}
\item \code{video.type}
\item \code{safe.search}
\item \code{event.type}
\item \code{channel.type}
}
}
\examples{
\dontrun{
# Authneticate
token <- youOauth(client.id = "something.apps.googleusercontent.com",
                  client.secret = "XxxXX1XxXxXxxx1xxx1xxXXX")

# find valid values for video.dimension
dims <- findParams(param = "video.dimension")

# set seed for reproducability
set.seed(19880525)

# fetch data using parameter
cats <- searchVideos(token, query = "cats", 
                     video.dimension = sample(dims, 1))
}

}
\author{
John Coene \email{jcoenep@hotmail.com}
}

