% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_parts.R
\name{findParts}
\alias{findParts}
\title{findParts}
\usage{
findParts(FUN)
}
\arguments{
\item{FUN}{Function to retrieve the valid \code{part}.}
}
\description{
Helper function to retrieve valid values for part 
(mostly used internally).
}
\details{
Valid \code{FUN}: 
\itemize{
\item \code{\link{getActivities}}
\item \code{\link{getCaptions}}
\item \code{\link{getChannels}}
\item \code{\link{getChannelSections}}
}
Alternatively see \code{scope} in \code{\link{youOAuth}}.
}
\examples{
\dontrun{
# Authenticate
token <- youOAuth(client.id = "something.apps.googleusercontent.com",
                  client.secret = "XxxXX1XxXxXxxx1xxx1xxXXX")
                  
# search channels about cats
search <- searchTube(token, query = "cats", type = "channel")

# random channel id
set.seed(19880525)
chan <- sample(search$id.channelId, 1)

# get valid part
p <- findParts(FUN = "getActivities")[1]

# fetch data
act <- getActivities(token, channel.id = chan, part = p)
}

}
\author{
\email{jcoenep@hotmail.com}
}

