% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_captions.R
\name{getCaptions}
\alias{getCaptions}
\title{getCaptions}
\usage{
getCaptions(token, video.id = NULL, part = "snippet", id,
  on.behalf.of.content.owner = NULL)
}
\arguments{
\item{token}{Your token as returned by \code{\link{youOAuth}}.}

\item{video.id}{specifies the YouTube video ID of the video for which the API should return 
caption tracks.}

\item{part}{specifies the caption resource parts that the API response will include. 
The default vlaue is \code{snippet}, can take any of \code{id} or 
\code{snippet}. See \code{link{findParts}}.}
}
\description{
Returns a list of caption tracks that are associated with a 
specified video.
}
\details{
Required authorisation: 
\itemize{
\item \code{force-ssl}
\item \code{partner-channel-audit}
}

See scope in \code{\link{youOAuth}}
}
\examples{
\dontrun{
# Authenticate
token <- youOauth(client.id = "something.apps.googleusercontent.com",
                  client.secret = "XxxXX1XxXxXxxx1xxx1xxXXX")
                  
# search videos about cats
search <- searchTube(token, query = "cats", type = "video")

# random channel id
set.seed(19880525)
vid <- sample(search$id.videoId, 1)

# fetch data
act <- getActivities(token, video.id = vid)
}

}
\author{
John Coene \email{jcoenep@hotmail.com}
}
\seealso{
\code{link{youOAuth}}, \code{\link{findParts}}
}

