% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{getCategories}
\alias{getCategories}
\title{getCategories}
\usage{
getCategories(token, id = NULL, region.code = NULL, hl = NULL)
}
\arguments{
\item{token}{Your token as returned by \code{\link{youOAuth}}.}

\item{id}{Specifies a comma-separated list of the YouTube channel category ID(s) for 
the resource(s) that are being retrieved. In a \code{guide.category} 
resource, the id property specifies the YouTube channel category ID.}

\item{region.code}{Instructs the API to return the list of guide categories available in the 
specified country. The parameter value is an 
\href{https://www.iso.org/obp/ui/#search}{ISO 3166-1 alpha-2} country code.}

\item{hl}{Specifies the language that will be used for text values in the API 
response. The default value (from YouTube) is \code{en-US}. See 
\code{\link{getLanguages}}}
}
\description{
Returns a list of categories that can be associated with 
YouTube channels.
}
\examples{
\dontrun{
token <- youOAuth(client.id = "something.apps.googleusercontent.com",
                  client.secret = "XxxXX1XxXxXxxx1xxx1xxXXX")

# fetch categories from Belgium
dull.cat <- getCategories(token, region.code = "BE")
}

}
\author{
John Coene \email{jcoenep@hotmail.com}
}

