% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel_sections.R
\name{getChannelSections}
\alias{getChannelSections}
\title{getChannelSections}
\usage{
getChannelSections(token, channel.id, part = "snippet", mine = FALSE, id,
  hl = NULL, on.behalf.of.content.owner = NULL, verbose = FALSE)
}
\arguments{
\item{token}{Your token as returned by \code{\link{youOAuth}}.}

\item{channel.id}{Indicates that the API response should only contain resources created by 
the channel. The default value is \code{NULL}.}

\item{part}{The part parameter specifies a comma-separated list of one or more activity 
resource properties that the API response will include. The default value 
is \code{snippet}, can take any of \code{contentDetails}, \code{id} or 
\code{snippet}. See \code{\link{findParts}}.}

\item{mine}{Set this parameter's value to true to retrieve a feed of the authenticated 
user's activities. The default value is \code{FALSE}.}

\item{id}{Specifies a comma-separated list of IDs that uniquely identify the 
channelSection resources that are being retrieved. In a 
\code{getChannelSections} resource, the id property specifies the section's 
ID.}

\item{hl}{The hl parameter instructs the API to retrieve localized resource metadata 
for a specific application language that the YouTube website supports. The 
parameter value must be a language code included in the list returned by 
\code{\link{getLanguages}}}

\item{on.behalf.of.content.owner}{Indicates that the request's 
authorization credentials identify a YouTube CMS user who is acting on 
behalf of the content owner specified in the parameter value. This parameter 
is intended for YouTube content partners that own and manage many different 
YouTube channels. It allows content owners to authenticate once and get 
access to all their video and channel data, without having to provide 
authentication credentials for each individual channel. The actual CMS 
account that the user authenticates with must be linked to the specified 
YouTube content owner. This parameter can only be used in a properly 
authorized request. Note: This parameter is intended exclusively for 
YouTube content partners. See scope under \code{\link{youOAuth}}. 
The default value is \code{NULL}.}

\item{verbose}{If \code{TRUE} prints infromational messages in the console. 
The default value is \code{FALSE}.}
}
\description{
Returns a list of channelSection resources that match the API 
request criteria.
}
\details{
Must specify one (and only one) of \code{mine} (\code{TRUE}), 
\code{id} or \code{channel.id}.
}
\author{
John Coene \email{jcoenep@hotmail.com}
}

