% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comment_threads.R
\name{getCommentThreads}
\alias{getCommentThreads}
\title{getCommentThreads}
\usage{
getCommentThreads(token, part = "snippet", n = 20, channel.id,
  all.threads.related.to.channel.id, id, video.id, max.results = NULL,
  moderation.status = "published", order = NULL, search.terms = NULL,
  text.format = NULL, verbose = FALSE)
}
\arguments{
\item{token}{Your token as returned by \code{\link{youOAuth}}.}

\item{part}{The part parameter specifies a comma-separated list of one or more activity 
resource properties that the API response will include. The default value 
is \code{snippet}, can take any of \code{snippet} or \code{id}. 
See \code{\link{findParts}}.}

\item{n}{Number of results to fecth. The default value is \code{20}.}

\item{channel.id}{Instructs the API to return comment threads containing comments about the 
specified channel. (The response will not include comments left on videos 
that the channel uploaded.).}

\item{all.threads.related.to.channel.id}{Instructs the API to return all comment threads associated with the 
specified channel. The response can include comments about the channel or 
about the channel's videos.}

\item{id}{Specifies a comma-separated list of comment thread IDs for the resources 
that should be retrieved.}

\item{video.id}{Instructs the API to return comment threads associated with the specified 
video ID.}

\item{max.results}{Specifies the maximum number of results that should be returned 
by each API call. Acceptable values are \code{0} to \code{100}, inclusive. 
The default value is \code{NULL}.}

\item{moderation.status}{Specifies the maximum number of items that should be returned in the result 
set. Note: This parameter is not supported for use in conjunction with the 
id parameter. Acceptable values are 1 to 100, inclusive. The default value 
is 20.}

\item{order}{Specifies the method that will be used to order resources in the API 
response. The default value is \code{NULL}. 
See \code{\link{findParams}} for valid values.}

\item{search.terms}{Instructs the API to limit the API response to only contain comments that 
contain the specified search terms.}

\item{text.format}{Set this parameter's value to html or plainText to instruct the API to 
return the comments left by users in html formatted or in plain text. 
The default value is \code{NULL} - which YouTube defaults to \code{html}.
See \code{\link{findParams}} for all valid values.}

\item{verbose}{If \code{TRUE} prints infromational messages in the console. 
The default value is \code{FALSE}.}
}
\description{
Returns a list of comment threads that match the API request 
parameters.
}
\details{
Must specify one (and only one) of 
\code{all.threads.related.to.channel.id} or \code{channel.id} or \code{id} or 
\code{video.id} 

The \code{part} parameter has great impact on your quota, visit the 
\href{https://developers.google.com/youtube/v3/getting-started#quota}{official documentation} 
for more information.
}
\examples{
\dontrun{
# Authenticate
token <- youOAuth(client.id = "something.apps.googleusercontent.com",
                  client.secret = "XxxXX1XxXxXxxx1xxx1xxXXX")

# search channel on cats
search <- searchTube(token, query = "cats", type = "channel")

# sample random channel
set.seed(19880525)
channel <- sample(search$snippet.id.channelId, 1)

# fetch comment thread
cats <- getCommentThreads(token, all.threads.related.to.channel.id = channel)
}

}
\author{
John Coene \email{jcoenep@hotmail.com}
}

