% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_items.R
\name{getPlaylistItems}
\alias{getPlaylistItems}
\title{getPlaylistItems}
\usage{
getPlaylistItems(token, part = "snippet", n = 50, id, playlist.id,
  max.results = 50, on.behalf.of.content.owner = NULL, video.id = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{Your token as returned by \code{\link{youOAuth}}.}

\item{part}{The part parameter specifies a comma-separated list of one or more activity 
resource properties that the API response will include. The default value 
is \code{snippet}, see \code{\link{findParts}} or 
\href{https://developers.google.com/youtube/v3/docs/channels/list}{official documentation} 
for all valid values.}

\item{n}{Number of results to fecth. The default value is \code{50}.}

\item{id}{The id parameter specifies a comma-separated list of one or more unique 
playlist item IDs.}

\item{playlist.id}{Specifies the unique ID of the playlist for which you want to retrieve 
playlist items. Note that even though this is an optional parameter, 
every request to retrieve playlist items must specify a value for either 
the id parameter or the \code{playlist.id} parameter.}

\item{max.results}{Specifies the maximum number of results that should be returned 
by each API call. Acceptable values are \code{0} to \code{50}, inclusive. 
The default value is \code{50}.}

\item{on.behalf.of.content.owner}{This parameter can only be used in a properly authorized request. Note: This 
parameter is intended exclusively for YouTube content partners. 
Indicates that the request's authorization credentials identify a YouTube 
CMS user who is acting on behalf of the content owner specified in the 
parameter value. This parameter is intended for YouTube content partners 
that own and manage many different YouTube channels. It allows content 
owners to authenticate once and get access to all their video and channel 
data, without having to provide authentication credentials for each 
individual channel. The CMS account that the user authenticates with must 
be linked to the specified YouTube content owner.}

\item{video.id}{Specifies that the request should return only the playlist items that 
contain the specified video.}

\item{verbose}{If \code{TRUE} prints infromational messages in the console. 
The default value is \code{FALSE}.}
}
\description{
Returns a collection of playlist items that match the API 
request parameters.
}
\details{
Must pass either \code{id} or \code{playlist.id}
}
\author{
John Coene \email{jcoenep@hotmail.com}
}

