% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region.R
\name{getRegions}
\alias{getRegions}
\title{getRegions}
\usage{
getRegions(token, hl = NULL, verbose = FALSE)
}
\arguments{
\item{token}{Your token as returned by \code{\link{youOAuth}}.}

\item{hl}{Specifies the language that should be used for text values in the API 
response. The default value is \code{en_US}. Optional argument.}

\item{verbose}{If \code{TRUE} prints infromational messages in the console. 
The default value is \code{FALSE}.}
}
\description{
Returns a list of application languages that the YouTube 
website supports.
}
\examples{
\dontrun{
# Authenticate
token <- youOAuth(client.id = "something.apps.googleusercontent.com",
                  client.secret = "XxxXX1XxXxXxxx1xxx1xxXXX")

# fetch languages
regions <- getRegions(token)
}

}
\author{
John Coene \email{jcoenep@hotmail.com}
}

