% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_categories.R
\name{getVideoCategories}
\alias{getVideoCategories}
\title{getVideoCategories}
\usage{
getVideoCategories(token, id, region.code, hl = "en_US")
}
\arguments{
\item{token}{Your token as returned by \code{\link{youOAuth}}.}

\item{id}{Specifies a comma-separated list of video category IDs for the resources 
that you are retrieving.}

\item{region.code}{Instructs the API to return the list of video categories available in the 
specified country. The parameter value is an 
\href{https://www.iso.org/obp/ui/#search}{ISO 3166-1 alpha-2} 
country code.}

\item{hl}{Specifies the language that should be used for text values in the API 
response. The default value is \code{en_US}.}
}
\description{
Returns a list of categories that can be associated with 
YouTube videos.
}
\details{
Must specify either \code{id} or \code{region.code}
}
\examples{
\dontrun{
# Authenticate
token <- youOAuth(client.id = "something.apps.googleusercontent.com",
                  client.secret = "XxxXX1XxXxXxxx1xxx1xxXXX")

# get categories
video.cat <- getVideoCategories(token, region.code = "US")
}

}
\author{
John Coene \email{jcoenep@hotmail.com}
}

