% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_videos.R
\name{getVideos}
\alias{getVideos}
\title{getVideos}
\usage{
getVideos(token, part = "snippet", n = 50, chart = "mostPopular", id,
  my.rating, hl = NULL, max.results = 50,
  on.behalf.of.content.owner = NULL, region.code = NULL,
  video.category.id = NULL, verbose = FALSE)
}
\arguments{
\item{token}{Your token as returned by \code{\link{youOAuth}}.}

\item{part}{The part parameter specifies a comma-separated list of one or more activity 
resource properties that the API response will include. The default value 
is \code{snippet}, see \code{\link{findParts}} for all valid values.}

\item{n}{Number of results to fecth. The default value is \code{50}.}

\item{chart}{The chart parameter identifies the chart that you want to retrieve. Only 
valid value is \code{mostPopular} or \code{NULL}. 
The default value if \code{mostPopular}.}

\item{id}{Specifies a comma-separated list of the YouTube video ID(s) for the 
resource(s) that are being retrieved. In a video resource, the id property 
specifies the video's ID.}

\item{my.rating}{This parameter can only be used in a properly authorized request. 
Set this parameter's value to \code{like} or \code{dislike} to instruct the 
API to only return videos liked or disliked by the authenticated user.}

\item{hl}{Specifies the language that will be used for text values in the API 
response. The default value (from YouTube) is \code{NULL}. See 
\code{\link{getLanguages}}. If localized resource details are available 
in that language, the resource's \code{snippet} (\code{part = "snippet"}).}

\item{max.results}{Specifies the maximum number of results that should be returned 
by each API call. Acceptable values are \code{0} to \code{50}, inclusive. 
The default value is \code{50}.}

\item{on.behalf.of.content.owner}{This parameter can only be used in a properly authorized request. Note: This 
parameter is intended exclusively for YouTube content partners.
Indicates that the request's authorization credentials identify a YouTube 
CMS user who is acting on behalf of the content owner specified in the 
parameter value. This parameter is intended for YouTube content partners 
that own and manage many different YouTube channels. It allows content 
owners to authenticate once and get access to all their video and channel 
data, without having to provide authentication credentials for each 
individual channel. The CMS account that the user authenticates with must be 
linked to the specified YouTube content owner.}

\item{region.code}{Instructs the API to select a video chart available in the specified region. 
This parameter can only be used in conjunction with the chart parameter. 
The parameter value is an ISO 3166-1 alpha-2 country code.}

\item{video.category.id}{Identifies the video category for which the chart should be retrieved. 
This parameter can only be used in conjunction with the chart parameter. 
By default, charts are not restricted to a particular category. The default 
value is 0.}

\item{verbose}{If \code{TRUE} prints infromational messages in the console. 
The default value is \code{FALSE}.}
}
\description{
Returns a list of videos that match the API request parameters.
}
\details{
Must specify one (and only one) of \code{chart}, \code{id} or 
\code{my.rating}
}
\examples{
\dontrun{
# Authenticate
token <- youOAuth(client.id = "something.apps.googleusercontent.com",
                  client.secret = "XxxXX1XxXxXxxx1xxx1xxXXX")

# get videos
videos <- getVideos(token)
}

}
\author{
John Coene \email{jcoenep@hotmail.com}
}

