% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youTubeDataR.R
\docType{package}
\name{youTubeDataR}
\alias{youTubeDataR}
\alias{youTubeDataR-package}
\title{youTubeDataR: R and Youtube.}
\description{
youTubeDataR integrates R and the YouTube Data API to allow fetching data from Youtube.
}
\section{youTubeDataR functions}{

youOAuth
findParams
findParts
getActivities
getCatptions
getChannelSections
getChannels
getCommentThreads
getLanguages
getComments
getGuideCategories
getRegion
getSubscriptions
getVideoCategories
getVideos
searchTube
}
\keyword{internal}

